增加empService介面跟實作
