+++ src/main/java/marketMaster/controller/employee/ChangePasswordController.java
@@ -8,14 +8,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import jakarta.servlet.http.HttpSession;
-import marketMaster.service.employee.EmployeeService;
+import marketMaster.service.employee.EmployeeServiceImpl;
 import marketMaster.viewModel.EmployeeViewModel;
 
 @Controller
 public class ChangePasswordController {
 
     @Autowired
-    private EmployeeService employeeService;
+    private EmployeeServiceImpl employeeService;
 
     @GetMapping("/employee/changePasswordPage")
     public String showChangePasswordPage() {
