+++ src/main/java/marketMaster/controller/employee/LoginController.java
@@ -9,14 +9,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import jakarta.servlet.http.HttpSession;
 import marketMaster.bean.employee.EmpBean;
-import marketMaster.service.employee.EmployeeService;
+import marketMaster.service.employee.EmployeeServiceImpl;
 import marketMaster.viewModel.EmployeeViewModel;
 
 @Controller
 public class LoginController {
 
 	@Autowired
-	private EmployeeService employeeService;
+	private EmployeeServiceImpl employeeService;
 	
 	@GetMapping("/employee/loginPage")
 	public String showLoginPage() {
