+++ src/main/java/marketMaster/service/restock/RestocksRepository.java
@@ -13,12 +13,12 @@
 
     public interface RestocksRepository extends JpaRepository<RestocksBean, String> {
         // 查詢最新的 Restock ID 並排序 mysql
-        @Query(value = "SELECT * FROM restocks WHERE restock_id LIKE CONCAT(:restockIdPattern, '%') ORDER BY restock_id DESC LIMIT 1", nativeQuery = true)
-        Optional<RestocksBean> findLatestRestockByDate(@Param("restockIdPattern") String restockIdPattern);
+//        @Query(value = "SELECT * FROM restocks WHERE restock_id LIKE CONCAT(:restockIdPattern, '%') ORDER BY restock_id DESC LIMIT 1", nativeQuery = true)
+//        Optional<RestocksBean> findLatestRestockByDate(@Param("restockIdPattern") String restockIdPattern);
 
 //        sqlServer
-//        @Query(value = "SELECT TOP 1 * FROM restocks WHERE restock_id LIKE :restockIdPattern + '%' ORDER BY restock_id DESC", nativeQuery = true)
-//        Optional<RestocksBean> findLatestRestockByDate(@Param("restockIdPattern") String restockIdPattern);
+        @Query(value = "SELECT TOP 1 * FROM restocks WHERE restock_id LIKE :restockIdPattern + '%' ORDER BY restock_id DESC", nativeQuery = true)
+        Optional<RestocksBean> findLatestRestockByDate(@Param("restockIdPattern") String restockIdPattern);
 
         //找到所有進貨Id並計算內明細的家總金額
         @Query("SELECT new marketMaster.DTO.restock.restock.RestockDTO( s.restockId,s.employee.employeeId,s.employee.employeeName,s.restockDate,s.restockTotalPrice)from RestocksBean s ")
